added app status service logic

+++ backend/src/main/java/com/jjimenez/healthservice/model/AppStatusDTO.java
@@ -1,4 +1,4 @@
 package com.jjimenez.healthservice.model;
 
-public record AppStatusDTO(String name, boolean running, Double uptimeSeconds) {
+public record AppStatusDTO(String name, boolean running, Long uptimeSeconds) {
 }
